requirements.txt
django=4.1.0
djangorestframework=3.14.0
djangorestframework-simplejwt=5.2.2
django-cors-headers=3.14.0
psycopg2-binary=2.9.6